backend/requirements.txt
fastapi>=0.104.1,<1.0.0
uvicorn>=0.24.0,<1.0.0
sqlalchemy>=2.0.23,<3.0.0
psycopg2-binary>=2.9.9,<3.0.0
pydantic>=2.5.0,<3.0.0
python-docx>=0.8.11,<1.0.0
PyPDF2>=3.0.1,<4.0.0
beautifulsoup4>=4.12.2,<5.0.0
requests>=2.31.0,<3.0.0
python-multipart>=0.0.6,<1.0.0
python-dateutil>=2.8.2,<3.0.0
numpy>=1.26.2,<2.0.0
pandas>=2.1.3,<3.0.0
pytest>=7.4.3,<9.0.0
pytest-asyncio>=0.21.1,<1.0.0
httpx>=0.25.2,<1.0.0
python-jose>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
email-validator>=2.1.0,<3.0.0
aiofiles>=23.2.1,<24.0.0
google-generativeai>=0.3.2,<1.0.0
groq>=0.4.2,<1.0.0
python-dotenv>=1.0.0,<2.0.0
gunicorn>=21.2.0,<23.0.0
openpyxl>=3.1.2,<4.0.0
